// frontend/src/app/models/user.ts
export class User {
  _id: string = '';
  username: string = '';
  password: string = '';
  safety_question: string = '';
  safety_answer: string = '';
  firstname: string = '';
  lastname: string = '';
  gender: string = '';
  address: string = '';
  phone: string = '';
  email: string = '';
  credit_card: string = '';
  profile_picture: string = '';
  type: string = '';
  restaurant: string = '';
  active: string = '';
  didnt_show_cnt: number = 0;
}